+++ NEWS.md
@@ -1,4 +1,4 @@
-# 0.2.0.9000
+# 0.3.0
 
 * Correct metadata upload (#55 - thanks AndrewMarritt)
 * Let `gcs_object_metaname` not require name so it can be reused (#56 - thanks seandavi)